requirements.txt
aioboto3==8.3.0
certifi==2020.12.5
chardet==3.0.4
googleapis-common-protos==1.6.0
greenlet==1.0.0
grpcio==1.33.2
idna==2.8
jsonschema==3.2.0
protobuf==3.12.4
PyAudio==0.2.11
PyMySQL==1.0.2
pytest-asyncio==0.14.0
python-dotenv==0.17.0
requests==2.22.0
six==1.15.0
SQLAlchemy==1.4.11
tinkoff-voicekit-client==0.3.0
urllib3==1.25.11
